# improve(finalizer): improve finalizer performance