--- a/src/dataworker/index.ts
+++ b/src/dataworker/index.ts
@@ -61,7 +61,6 @@ export async function runDataworker(_logger: winston.Logger, baseSigner: Signer)
     message: `Time to update non-spoke clients: ${(performance.now() - loopStart) / 1000}s`,
   });
   loopStart = performance.now();
-
   let bundleDataToPersist: BundleDataToPersistToDALayerType | undefined = undefined;
   try {
     logger[startupLogLevel(config)]({ at: "Dataworker#index", message: "Dataworker started 👩‍🔬", config });
@@ -108,7 +107,7 @@ export async function runDataworker(_logger: winston.Logger, baseSigner: Signer)
         fromBlocks,
         toBlocks
       );
-
+      const dataworkerFunctionLoopTimerStart = performance.now();
       // Validate and dispute pending proposal before proposing a new one
       if (config.disputerEnabled) {
         await dataworker.validatePendingRootBundle(spokePoolClients, config.sendingDisputesEnabled, fromBlocks);
@@ -185,11 +184,14 @@ export async function runDataworker(_logger: winston.Logger, baseSigner: Signer)
         });
       }
 
+      const dataworkerFunctionLoopTimerEnd = performance.now();
       logger.debug({
         at: "Dataworker#index",
         message: `Time to update spoke pool clients and run dataworker function: ${
           (performance.now() - loopStart) / 1000
         }s`,
+        timeToLoadSpokes: (dataworkerFunctionLoopTimerStart - loopStart) / 1000,
+        timeToRunDataworkerFunctions: (dataworkerFunctionLoopTimerEnd - dataworkerFunctionLoopTimerStart) / 1000,
       });
       loopStart = performance.now();