--- a/src/finalizer/index.ts
+++ b/src/finalizer/index.ts
@@ -413,9 +413,9 @@ export async function runFinalizer(_logger: winston.Logger, baseSigner: Signer):
 
       logger.debug({
         at: "Finalizer#index",
-        message: `Time to loop: ${(loopEndPostFinalizations - loopStart) / 1000}s`,
-        timeToUpdateSpokeClients: (loopStartPostSpokePoolUpdates - loopStart) / 1000,
-        timeToFinalize: (loopEndPostFinalizations - loopStartPostSpokePoolUpdates) / 1000,
+        message: `Time to loop: ${Math.round((loopEndPostFinalizations - loopStart) / 1000)}s`,
+        timeToUpdateSpokeClients: Math.round((loopStartPostSpokePoolUpdates - loopStart) / 1000),
+        timeToFinalize: Math.round((loopEndPostFinalizations - loopStartPostSpokePoolUpdates) / 1000),
       });
 
       if (await processEndPollingLoop(logger, "Dataworker", config.pollingDelay)) {