This PR aims to parallelize the finalizer code using `Promise.all( )`. Instrumentation that breaks down speed of spokes & finalizer/dataworker functions has also been added to capture improvements of this change. 

A draw of this PR is that it makes logs produced by the finalizer less linear.

Timings **prior** to optimization

```
2024-04-12 17:46:24 [debug]: {
  "at": "Finalizer#index",
  "message": "Time to loop: 85.80977033400535s",
  "timeToUpdateSpokeClients": 23.340696208953858,
  "timeToFinalize": 62.4690741250515,
  "bot-identifier": "not-james"
}

2024-04-12 17:49:17 [debug]: {
  "at": "Finalizer#index",
  "message": "Time to loop: 85.8968901669979s",
  "timeToUpdateSpokeClients": 23.311546999931334,
  "timeToFinalize": 62.585343167066576,
  "bot-identifier": "not-james"
}
```

Timings **post** optimization

```
2024-04-12 17:40:28 [debug]: {
  "at": "Finalizer#index",
  "message": "Time to loop: 67.65214512503147s",
  "timeToUpdateSpokeClients": 31.18309954202175,
  "timeToFinalize": 36.469045583009716,
  "bot-identifier": "not-james"
}

2024-04-12 17:51:23 [debug]: {
  "at": "Finalizer#index",
  "message": "Time to loop: 58.301649749994276s",
  "timeToUpdateSpokeClients": 24.055480499982835,
  "timeToFinalize": 34.24616925001144,
  "bot-identifier": "not-james"
}
```